fix(ui): isolate Pi skill display helper
