+++ packages/ui/src/features/sessions/components/session-update/parseFileMentions.tsx
@@ -16,19 +16,6 @@ const MENTION_TAG_REGEX =
 const MENTION_TAG_TEST =
   /<(?:file\s+path|folder\s+path|github_issue\s+number|github_pr\s+number|error_context\s+label)="[^"]+"/;
 const SLASH_COMMAND_START = /^\/([a-zA-Z][\w-]*)(?=\s|$)/;
-const PI_SKILL_INVOCATION =
-  /^<skill name="([^"]+)" location="[^"]+">\n[\s\S]*?\n<\/skill>(?:\n\n([\s\S]+))?$/;
-
-export function collapsePiSkillInvocation(content: string): string {
-  const match = content.match(PI_SKILL_INVOCATION);
-  if (!match) {
-    return content;
-  }
-
-  const name = unescapeXmlAttr(match[1]);
-  const userMessage = match[2]?.trim();
-  return userMessage ? `/${name}\n\n${userMessage}` : `/${name}`;
-}
 
 const inlineComponents: Components = {
   ...baseComponents,
